requirements.txt
pandas~=2.0.2
xgboost~=1.7.5
Django~=4.2.0
requests~=2.31.0
celery~=5.2.7
django-celery-results~=2.5.1
django-celery-beat~=2.5.0
djangorestframework~=3.15.1
playwright~=1.49.1
fastapi~=0.115.6
uvicorn~=0.22.0
redis~=5.0.0
hiredis~=2.0.0
